requirements.txt
streamlit>=1.28.0
pydub>=0.25.1
openai>=1.0.0
python-dotenv>=1.0.0